removing dependency on drupal user system
I think this simplifies the system greatly by dropping the hooks into the drupal user system. The advantage is that it's far easier to add and edit team members, and allows for a person to enter multiple entries for a person (more than one cell number). The disadvantage is only present if the potential team members are already deeply entrenched into a drupal user system.